+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ﻿# Changelog
 
+## 1.54.24
+
+- Added filename to DhcpOptionType
+
+## 1.54.23
+
+- Added next_server to DhcpOptionType
+
+## 1.54.22
+
+- Added undocumented boot_options_enabled to DhcpOptionType
+
 ## 1.54.12
 
 - WPA 192-bit Security added to appliance SSID